+++ NEWS.md
@@ -1,5 +1,10 @@
-## formatters 0.5.1.9001
+## formatters 0.5.1.9002
+ * Added decimal alignment support with `decimal`, `dec_right`, and `dec_left`.
  * Updated `export_as_txt` to pass `rep_cols` arugment to pagination function.
+
+### Miscellaneous
+ * Added `list_valid_aligns()` that lists the available alignments.
+
 ## formatters 0.5.1
  * Updated `export_as_txt` to print split label when using `page_by` with only one level.
  * Reduced the amount of spelling issues.